Add files via upload
Assets are the major area to generate revenue for any financial institution. The project is all about prediction whether a particular borrower will default while paying the loan or not and hence minimizing the risk factor for the financial institution. The dataset is divided into test and train data set and final prediction is stored in a csv file.